fix FileSystem Export and ExportSet discovery
