+++ visualiser/discovery/models/object_storage.py
@@ -1,24 +1,6 @@
 # Copyright © 2021, Oracle and/or its affiliates. All rights reserved.
 
-from oci.object_storage.models import BucketSummary, PreauthenticatedRequestSummary
-
-# extend BucketSummary to add compartment id
-class ExtendedBucketSummary(BucketSummary):
-
-    _compartment_id = None
-
-    @property
-    def compartment_id(self):
-        return self._compartment_id
-
-    def __init__(self, compartment_id, bucket: BucketSummary):
-        attrs = [item for item in dir(BucketSummary) if not item.startswith('_') and item[0].islower()]
-        init_args = {}
-        for attr in attrs:
-            init_args[attr] = getattr(bucket, attr)
-        super().__init__(**init_args)
-        self._compartment_id = compartment_id
-        self.swagger_types.update({"compartment_id": "str"})
+from oci.object_storage.models import PreauthenticatedRequestSummary
 
 # extend PreauthenticatedRequest to add compartment id  and bucket id
